+++ Media Tagging Manager/ServiceRegistrator.cs
@@ -14,13 +14,15 @@ public sealed class ServiceRegistrator : IPluginServiceRegistrator
     {
         serviceCollection.AddSingleton<ScanStateStore>();
         serviceCollection.AddSingleton<TagBackupManager>();
+        serviceCollection.AddSingleton<WatchmodeQuotaTracker>();
         serviceCollection.AddSingleton<ProviderNetworkScanner>();
         serviceCollection.AddSingleton<ManualScanRequestQueue>();
         serviceCollection.AddSingleton<MediaBrowser.Model.Tasks.IScheduledTask, RefreshAvailabilityTask>();
         serviceCollection.AddSingleton<MediaBrowser.Model.Tasks.IScheduledTask, ManualScanTask>();
         serviceCollection.AddSingleton<MediaBrowser.Controller.Library.ILibraryPostScanTask, NewMediaPostScanTask>();
-        serviceCollection.AddHttpClient<IAvailabilitySource, TmdbAvailabilitySource>();
-        serviceCollection.AddHttpClient<IAvailabilitySource, WatchmodeAvailabilitySource>();
-        serviceCollection.AddHttpClient<IAvailabilitySource, CustomJsonAvailabilitySource>();
+        serviceCollection.AddHttpClient<TmdbAvailabilitySource>();
+        serviceCollection.AddHttpClient<WatchmodeAvailabilitySource>();
+        serviceCollection.AddSingleton<IAvailabilitySource>(serviceProvider => serviceProvider.GetRequiredService<TmdbAvailabilitySource>());
+        serviceCollection.AddSingleton<IAvailabilitySource>(serviceProvider => serviceProvider.GetRequiredService<WatchmodeAvailabilitySource>());
     }
 }